Moving selector into query package
