+++ zenoh-kotlin/src/commonMain/kotlin/io/zenoh/Session.kt
@@ -33,7 +33,7 @@ import io.zenoh.query.*
 import io.zenoh.query.Query
 import io.zenoh.query.Queryable
 import io.zenoh.sample.Sample
-import io.zenoh.selector.Selector
+import io.zenoh.query.Selector
 import io.zenoh.qos.Reliability
 import io.zenoh.session.SessionDeclaration
 import io.zenoh.session.SessionInfo
