+++ zenoh-kotlin/src/commonMain/kotlin/io/zenoh/query/Parameters.kt
@@ -12,7 +12,7 @@
 //   ZettaScale Zenoh Team, <zenoh@zettascale.tech>
 //
 
-package io.zenoh.selector
+package io.zenoh.query
 
 import java.net.URLDecoder
