+++ zenoh-kotlin/src/commonTest/kotlin/io/zenoh/GetTest.kt
@@ -20,9 +20,9 @@ import io.zenoh.bytes.into
 import io.zenoh.config.Config
 import io.zenoh.query.Reply
 import io.zenoh.query.Queryable
-import io.zenoh.selector.Parameters
-import io.zenoh.selector.Selector
-import io.zenoh.selector.intoSelector
+import io.zenoh.query.Parameters
+import io.zenoh.query.Selector
+import io.zenoh.query.intoSelector
 import org.apache.commons.net.ntp.TimeStamp
 import java.time.Duration
 import java.util.*
